requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
